# Dev Feature 8 - Contract Wide CPU Optimizations