--- a/include/checkformat.hpp
+++ b/include/checkformat.hpp
@@ -29,7 +29,7 @@ For a format to be valid, three things are checked:
 Note: This could all be done a lot cleaner by using regex or similar libraries
       However, using them would bloat up the contract size significantly.
 */
-void check_format(vector <FORMAT> lines) {
+void check_format(vector <FORMAT> & lines) {
 
     bool found_name = false;