--- a/include/checkformat.hpp
+++ b/include/checkformat.hpp
@@ -35,10 +35,10 @@ void check_format(vector <FORMAT> & lines) {
 
     vector <string> attribute_names = {};
 
-    for (FORMAT line : lines) {
+    for (FORMAT & line : lines) {
 
-        string name = line.name;
-        string type = line.type;
+        string & name = line.name;
+        string & type = line.type;
 
         check(name.length() != 0, "An attribute's name can't be empty");
         check(name.length() <= 64, "An attribute's name can only be 64 characters max");