Note: This merge is for demonstration purposes & will be rewritten as a final form once all of the other features are implemented.

Overall improvements are benchmarked with minting 100x templated assets, with no mutable or immutable data & 4 schema fields, resulting in:
- Current version: 4.25ms to 8.54ms range, average of 6.385ms
- Upgraded version: 3ms to 7ms range, average of 5ms
- Estimated 1ms to 1.5ms range of reduction in computation time
- Estimated 20% to 30% range of CPU cost reduction at the high end
- Estimated 4% to 7% range of CPU cost reduction in practice (due to more data being in mint asset actions, more schema fields, smaller collection's 'serialized_data')

**Table Fetch Improvements**
When a table fetch is defined as so:
```C++
config_t config = config_t(get_self(), get_self().value);
```

It makes it so that the table is initialized & stored in memory on every single action call to the contract (including inline actions to itself, such as the logs). Essentially, something like 10x tables are being initialized every single time the contract executes any action, adding an unnecessary overhead to CPU costs. 

They have been rewritten to be as so:
```C++
config_t get_config {return config_t(get_self(), get_self().value);}
```

Making it so that they are initialized only when the function is called instead of being implicitly initialized, helping mitigate CPU costs. All table fetches have been redefine according to this format.

**Collection Data Improvements**
All collections' table fetches are consuming (up to) an extra 3kb of on-chain data stored under the 'serialized_data' field, which no smart contract actually needs to interact with. This data is then fetched again during logs, making it so that every 'mintasset' & 'setassetdata' action is fetching (up to) 6kb of unnecessary data. 

This upgrade moves the 'serialized_data' to the **'coldata'** scope using the same primary_key, as well as implementing automatic conversions to this new approach upon any interaction with a collections' fees, authorized_accounts, notify_accounts or serialized_data. 

Further more, 'check_has_collection_auth' has been rewritten to be more CPU performative by not performing a "find" operation on the collections table twice by passing the collection_itr pointer to the function directly